ReFact: 주석 수정, assets file 수정
